nic_drv: remove alignment warning
The warning popped up due to changes from genodelabs/genode#4312. On a
second thought, and validation with the specs as well as test-run on
hardware, it is safe to remove the warning. According to the ug585,
there might be a impact on the performance due to the new alignment, yet
this is unlikely to be noticeable (see comment in code).