fpga_drv: fix occasional freeze on reconfiguration
Reconfiguration apparently also requires asserting the FPGA resets and
setting the voltage level shifters. This is now supported by the
platform driver, which provides the reset domain "fpga_nreset" for this
purpose. By binding this reset domain to the devcfg device, the fpga
driver is able to halt the fpga when it acquires the device and to reset
the fpga when it releases the device.